+++ presto-main/src/test/java/com/facebook/presto/operator/window/TestPercentRankFunction.java
@@ -18,6 +18,7 @@ import org.testng.annotations.Test;
 import static com.facebook.presto.SessionTestUtils.TEST_SESSION;
 import static com.facebook.presto.spi.type.BigintType.BIGINT;
 import static com.facebook.presto.spi.type.DoubleType.DOUBLE;
+import static com.facebook.presto.spi.type.IntegerType.INTEGER;
 import static com.facebook.presto.spi.type.VarcharType.VARCHAR;
 import static com.facebook.presto.testing.MaterializedResult.resultBuilder;
 
@@ -28,17 +29,17 @@ public class TestPercentRankFunction
     public void testPercentRank()
     {
         assertWindowQuery("percent_rank() OVER (PARTITION BY orderstatus ORDER BY orderkey)",
-                resultBuilder(TEST_SESSION, BIGINT, VARCHAR, DOUBLE)
-                        .row(3L, "F", 0.0)
-                        .row(5L, "F", 1 / 3.0)
-                        .row(6L, "F", 2 / 3.0)
-                        .row(33L, "F", 1.0)
-                        .row(1L, "O", 0.0)
-                        .row(2L, "O", 0.2)
-                        .row(4L, "O", 0.4)
-                        .row(7L, "O", 0.6)
-                        .row(32L, "O", 0.8)
-                        .row(34L, "O", 1.0)
+                resultBuilder(TEST_SESSION, INTEGER, VARCHAR, DOUBLE)
+                        .row(3, "F", 0.0)
+                        .row(5, "F", 1 / 3.0)
+                        .row(6, "F", 2 / 3.0)
+                        .row(33, "F", 1.0)
+                        .row(1, "O", 0.0)
+                        .row(2, "O", 0.2)
+                        .row(4, "O", 0.4)
+                        .row(7, "O", 0.6)
+                        .row(32, "O", 0.8)
+                        .row(34, "O", 1.0)
                         .build());
         assertWindowQueryWithNulls("percent_rank() OVER (PARTITION BY orderstatus ORDER BY orderkey)",
                 resultBuilder(TEST_SESSION, BIGINT, VARCHAR, BIGINT)
@@ -55,17 +56,17 @@ public class TestPercentRankFunction
                         .build());
 
         assertWindowQuery("percent_rank() OVER (ORDER BY orderkey)",
-                resultBuilder(TEST_SESSION, BIGINT, VARCHAR, DOUBLE)
-                        .row(1L, "O", 0.0)
-                        .row(2L, "O", 1 / 9.0)
-                        .row(3L, "F", 2 / 9.0)
-                        .row(4L, "O", 3 / 9.0)
-                        .row(5L, "F", 4 / 9.0)
-                        .row(6L, "F", 5 / 9.0)
-                        .row(7L, "O", 6 / 9.0)
-                        .row(32L, "O", 7 / 9.0)
-                        .row(33L, "F", 8 / 9.0)
-                        .row(34L, "O", 1.0)
+                resultBuilder(TEST_SESSION, INTEGER, VARCHAR, DOUBLE)
+                        .row(1, "O", 0.0)
+                        .row(2, "O", 1 / 9.0)
+                        .row(3, "F", 2 / 9.0)
+                        .row(4, "O", 3 / 9.0)
+                        .row(5, "F", 4 / 9.0)
+                        .row(6, "F", 5 / 9.0)
+                        .row(7, "O", 6 / 9.0)
+                        .row(32, "O", 7 / 9.0)
+                        .row(33, "F", 8 / 9.0)
+                        .row(34, "O", 1.0)
                         .build());
         assertWindowQueryWithNulls("percent_rank() OVER (ORDER BY orderkey)",
                 resultBuilder(TEST_SESSION, BIGINT, VARCHAR, BIGINT)
@@ -82,17 +83,17 @@ public class TestPercentRankFunction
                         .build());
 
         assertWindowQuery("percent_rank() OVER (ORDER BY orderstatus)",
-                resultBuilder(TEST_SESSION, BIGINT, VARCHAR, DOUBLE)
-                        .row(3L, "F", 0.0)
-                        .row(5L, "F", 0.0)
-                        .row(6L, "F", 0.0)
-                        .row(33L, "F", 0.0)
-                        .row(1L, "O", 4 / 9.0)
-                        .row(2L, "O", 4 / 9.0)
-                        .row(4L, "O", 4 / 9.0)
-                        .row(7L, "O", 4 / 9.0)
-                        .row(32L, "O", 4 / 9.0)
-                        .row(34L, "O", 4 / 9.0)
+                resultBuilder(TEST_SESSION, INTEGER, VARCHAR, DOUBLE)
+                        .row(3, "F", 0.0)
+                        .row(5, "F", 0.0)
+                        .row(6, "F", 0.0)
+                        .row(33, "F", 0.0)
+                        .row(1, "O", 4 / 9.0)
+                        .row(2, "O", 4 / 9.0)
+                        .row(4, "O", 4 / 9.0)
+                        .row(7, "O", 4 / 9.0)
+                        .row(32, "O", 4 / 9.0)
+                        .row(34, "O", 4 / 9.0)
                         .build());
         assertWindowQueryWithNulls("percent_rank() OVER (ORDER BY orderstatus)",
                 resultBuilder(TEST_SESSION, BIGINT, VARCHAR, BIGINT)
@@ -109,17 +110,17 @@ public class TestPercentRankFunction
                         .build());
 
         assertWindowQuery("percent_rank() OVER (PARTITION BY orderkey)",
-                resultBuilder(TEST_SESSION, BIGINT, VARCHAR, DOUBLE)
-                        .row(1L, "O", 0.0)
-                        .row(2L, "O", 0.0)
-                        .row(3L, "F", 0.0)
-                        .row(4L, "O", 0.0)
-                        .row(5L, "F", 0.0)
-                        .row(6L, "F", 0.0)
-                        .row(7L, "O", 0.0)
-                        .row(32L, "O", 0.0)
-                        .row(33L, "F", 0.0)
-                        .row(34L, "O", 0.0)
+                resultBuilder(TEST_SESSION, INTEGER, VARCHAR, DOUBLE)
+                        .row(1, "O", 0.0)
+                        .row(2, "O", 0.0)
+                        .row(3, "F", 0.0)
+                        .row(4, "O", 0.0)
+                        .row(5, "F", 0.0)
+                        .row(6, "F", 0.0)
+                        .row(7, "O", 0.0)
+                        .row(32, "O", 0.0)
+                        .row(33, "F", 0.0)
+                        .row(34, "O", 0.0)
                         .build());
         assertWindowQueryWithNulls("percent_rank() OVER (PARTITION BY orderkey)",
                 resultBuilder(TEST_SESSION, BIGINT, VARCHAR, BIGINT)
